Refactor: Modularized notificationService.js into Template, Transport, and Preference services (#108)
